manage extra svcs separately, auto-enable running
Add service_imdc_manage and service_imflows_manage which allow the user
to turn off management of individual extra services. This is easier than
maintaining a whitelist of versions of intermapper that ship with
these services.

Any service that is set to ensure running or true will automatically
have it's enable property set to true, so that the service will start on
boot.